Move assign_impl to assign_helpers
